vblank: set SCHED_RR for the SGI_video_sync thread
