+++ src/utils.c
@@ -1,8 +1,10 @@
+#include <pthread.h>
 #include <stdio.h>
 #include <string.h>
 #include <sys/uio.h>
 
 #include "compiler.h"
+#include "rtkit.h"
 #include "string_utils.h"
 #include "test.h"
 #include "utils.h"
@@ -270,6 +272,41 @@ void rolling_quantile_pop_front(struct rolling_quantile *rq, int x) {
 	if (x <= rq->estimate) {
 		rq->current_rank--;
 	}
+}
+
+/// Switch to real-time scheduling policy (SCHED_RR) if possible
+///
+/// Make picom realtime to reduce latency, and make rendering times more predictable to
+/// help pacing.
+///
+/// This requires the user to set up permissions for the real-time scheduling. e.g. by
+/// setting `ulimit -r`, or giving us the CAP_SYS_NICE capability.
+void set_rr_scheduling(void) {
+	int priority = sched_get_priority_min(SCHED_RR);
+
+	if (rtkit_make_realtime(0, priority)) {
+		log_info("Set realtime priority to %d with rtkit.", priority);
+		return;
+	}
+
+	// Fallback to use pthread_setschedparam
+	struct sched_param param;
+	int old_policy;
+	int ret = pthread_getschedparam(pthread_self(), &old_policy, &param);
+	if (ret != 0) {
+		log_info("Couldn't get old scheduling priority.");
+		return;
+	}
+
+	param.sched_priority = priority;
+
+	ret = pthread_setschedparam(pthread_self(), SCHED_RR, &param);
+	if (ret != 0) {
+		log_info("Couldn't set real-time scheduling priority to %d.", priority);
+		return;
+	}
+
+	log_info("Set real-time scheduling priority to %d.", priority);
 }
 
 // vim: set noet sw=8 ts=8 :
